src/requirements.txt
google-cloud-secret-manager
google-cloud-storage
requests